more for #128, #129 new tests required
