+++ storage/backend/routes/file.js
@@ -4,23 +4,22 @@ var {validateBody} = require("lib")
 const app = express.Router()
 
 const {
-	validateCameraAndAppendToPath, 
-	directoryList, 
-	deleteFileDirectory, 
-	validateDays, 
-	filterList,
+	validateCameraAndAppendToPath,
+	validateDays,
+	getCameraMetricFromDatabase,
+	updateDeletionOfFiles,
+	deleteFileDirectory,
+	deleteFilesBeforeDateGlob,
 	fileStats,
-	getCachedFileData,
-	deleteFilesBeforeDateGlob, 
-} = require("./lib/file.js")
+	cameraMetrics
+} = require('./lib/file.js');
+
+app.post("/pathSize", validateBody, validateCameraAndAppendToPath, getCameraMetricFromDatabase('size')) 
+app.post("/pathFileCount", validateBody, validateCameraAndAppendToPath, getCameraMetricFromDatabase('count'))
+app.post("/pathDelete", validateBody, validateCameraAndAppendToPath, updateDeletionOfFiles('directory'), deleteFileDirectory) 
+app.post("/pathClean", validateBody, validateCameraAndAppendToPath, validateDays, updateDeletionOfFiles('files'), deleteFilesBeforeDateGlob)
 
 app.get("/pathStats", fileStats)
-
-app.post("/pathSize", validateBody, validateCameraAndAppendToPath, getCachedFileData("size")) 
-app.post("/pathFileCount", validateBody, validateCameraAndAppendToPath, getCachedFileData("count"))
-app.post("/pathDelete", validateBody, validateCameraAndAppendToPath, deleteFileDirectory) 
-app.post("/pathClean", validateBody, validateCameraAndAppendToPath, validateDays, directoryList, filterList("before"), deleteFilesBeforeDateGlob)
-
-app.post("/pathMetrics", getCachedFileData("all"))
+app.post("/pathMetrics", cameraMetrics)
 
 module.exports = app
